Refactored file systems for easy of use
