+++ src/service.ts
@@ -1,10 +1,8 @@
 export default {
-    hello: data => {
-        return 'HELLO - ' + data.name + " " + data.age;
-    },
     addLocation: data => {
         return {
-            message: 'Location Added: ID: ' + data.id + ', cityName: ' + data.cityName + ', planetName: ' + data.planetName + ".",
+            message: 'Location Added: ID: ' + data.id + ', city name: ' + data.cityName + ', planet name: ' + data.planetName
+                + ", total capacity: " + data.totalAvailableCapacity + ", curent available space: " + data.currentAvailableSpace,
         }
     }
 }
